[agent_builder/observability/charts] - refactor: streamline SkillUsageChart tooltip rendering
- Removed separate `SkillSourceTooltip` component, integrating its functionality directly into `SkillUsageChart`.
- Simplified tooltip code by using `ChartTooltipCard` and eliminating redundant checks and mappings.
- Dropped unused imports following the tooltip refactoring to clean up dependencies.
